Change API to use files, not strings for checking
Prior to this commit the main API was to send strings in and get your
results back, but that wasn't really that helpful. As soon as I started
trying to use the gem to script something, I realized that what I really
wanted was to pass in files and get back my results that way.

This commit merely changes the API to take in files, not strings.