Fix column coordinate
`hunspell` gives you what it refers to as the offset of the result, but
that's not helpful - what you want are the coordinates that will get you
to the word. In Vim, I want to be able to do 10G, 10| or whatever and
jump right to the damn position, not the character BEFORE the mistake.