Don't spell check inline code
Whew! This was a hard one. I wanted to keep things in `sed` but I
couldn't figure out a way to get the length of the match and replace
that many 1s. I asked Josh and Jack for advice and Jack's was to use
Ruby, haha.

I feel like that one-liner, while cryptic, isn't too bad and I did leave
a comment about what it's doing.

I may end up pulling all these things that transform the incoming
document into a script or something in the repo so that it's a little
easier to reason about, but this works for now.

Closes #5